+++ src/components/index.ts
@@ -1,3 +1,4 @@
-export { default as cVasDemoCard } from './c-vas-demo-card.vue';
-export { default as cVasDemoTable } from './c-vas-demo-table.vue';
-export { default as cVasRenderMarkdownFile } from './c-vas-render-markdown-file.vue';
+// DEPRECATED
+export { default as cVasDemoCard } from '../features/c-vas-demo-card.vue';
+export { default as cVasDemoTable } from '../featuresc-vas-demo-table.vue';
+export { default as cVasRenderMarkdownFile } from '../featuresc-vas-render-markdown-file.vue';
