+++ src/features/index.ts
@@ -1,0 +1,3 @@
+export { default as cVasDemoCard } from './c-vas-demo-card.vue';
+export { default as cVasDemoTable } from './c-vas-demo-table.vue';
+export { default as cVasRenderMarkdownFile } from './c-vas-render-markdown-file.vue';
