+++ src/styleguide/demo-pages/elements/r-vas-radio.vue
@@ -35,7 +35,7 @@
 
 <script lang="ts">
   import { defineComponent } from 'vue';
-  import cVasDemoCard from '../../../components/c-vas-demo-card.vue';
+  import cVasDemoCard from '../../../features/c-vas-demo-card.vue';
   import eVasCheckbox from '../../../elements/e-vas-checkbox.vue';
   import eVasRadio from '../../../elements/e-vas-radio.vue';
   import lVasLayout from '../../../layouts/l-vas-layout.vue';
